moved filters components to dir
